+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `send` methods to the `SolidClient` API to allow testers to send requests with full control over the headers and
   the ability to use methods not defined by the HTTP specification.
 * Push comments (for features, scenarios and steps) found in the test files through to the results reports.
+* Use tags in feature files to allow tests to be skipped based on the server definition
 
 ### Minor changes
 * Shorten the generated names of test resources.